+++ frontend/src/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './bills';
 export * from './home';
+export * from './income';